requirements.txt
alembic==1.12.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
certifi==2023.11.17
click==8.1.7
cloudinary==1.39.0
distro==1.9.0
exceptiongroup==1.2.2
flasgger==0.9.5
Flask==2.2.5
Flask-Admin==1.6.1
Flask-Cors==3.0.10
Flask-Migrate==4.0.5
Flask-SQLAlchemy==2.5.1
flask-swagger-ui==4.11.1
greenlet==3.1.1
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Mako==1.3.2
MarkupSafe==3.0.2
mistune==3.1.3
openai==0.28.1
packaging==24.2
psycopg2-binary==2.9.10
pydantic==2.11.2
pydantic_core==2.33.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-editor==1.0.4
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.24.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==1.4.49
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.1
urllib3==2.2.1
Werkzeug==2.2.3
WTForms==3.1.2
flask-jwt-extended==4.6.0

